feat(upload): rename enclave_actor.rs
