+++ rtc_data_service/src/lib.rs
@@ -4,6 +4,6 @@
 #![warn(rust_2018_idioms)]
 
 pub mod app_config;
-pub mod enclave_actor;
+pub mod data_enclave_actor;
 pub mod handlers;
 pub mod merge_error;
